requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
fake-useragent==0.1.11
future==0.17.1
Google-Search-API==1.1.14
idna==2.8
multidict==4.5.2
nltk==3.4
numpy==1.16.1
pandas==0.24.0
pkg-resources==0.0.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
selenium==2.53.6
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.7.3
Unidecode==1.0.23
urllib3==1.24.1
vcrpy==2.0.1
wrapt==1.11.1
yarl==1.3.0
